AI-Model-JVFA/requirements.txt
Flask
flask-cors
python-dotenv
openai
langchain
chromadb
tiktoken
unstructured[all-docs]